[plib] Compacts metadata by assuming fixed size of log entry.

+++ plib/async_file_store.h
@@ -23,9 +23,9 @@ class AsyncFileStore : public FileStore {
   AsyncFileStore(const char *name, int num_out, int num_in) :
       FileStore(name, num_out, num_in) { }
 
-  void *Submit(const DataEntry data[], int n);
+  void *Submit(const DataEntry data[], uint32_t n);
   int Commit(void *handle, uint64_t timestamp,
-      const MetaEntry meta[], int16_t n);
+      const MetaEntry meta[], uint32_t n);
  private:
   void AioWrite(aiocb *cb, File &file,
       void *buf, size_t nbytes, int priority = 0);
